% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-utils.R
\name{as_float_tensor}
\alias{as_float_tensor}
\title{Convert an R matrix / data.frame to a float torch tensor}
\usage{
as_float_tensor(x, device = "cpu")
}
\arguments{
\item{x}{Numeric matrix or data.frame.}

\item{device}{Character device string.}
}
\value{
A 2-D \code{torch_tensor} of \code{torch_float()}.
}
\description{
Convert an R matrix / data.frame to a float torch tensor
}
\keyword{internal}
