% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-bar.R
\name{bar_logits_to_quantiles}
\alias{bar_logits_to_quantiles}
\title{Convert bar-distribution logits to requested quantiles}
\usage{
bar_logits_to_quantiles(logits, borders, quantiles)
}
\arguments{
\item{logits}{\code{torch_tensor} of shape \verb{(..., n_bar_bins)}.}

\item{borders}{\code{torch_tensor} of length \code{n_bar_bins + 1}.}

\item{quantiles}{Numeric vector of quantiles in \verb{(0, 1)}.}
}
\value{
\code{torch_tensor} with the quantile axis appended last.
}
\description{
Convert bar-distribution logits to requested quantiles
}
