% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep-transforms.R
\name{.svd_n_components}
\alias{.svd_n_components}
\title{Compute n_components for the svd transformer preset.}
\usage{
.svd_n_components(global_name, n_samples, n_features)
}
\description{
Compute n_components for the svd transformer preset.
}
\keyword{internal}
