% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-mitra.R
\name{mitra_y_embedding_regression}
\alias{mitra_y_embedding_regression}
\title{Target embedding for scalar regression}
\usage{
mitra_y_embedding_regression(dim)
}
\description{
Target embedding for scalar regression
}
\keyword{internal}
