% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-tabpfn-encoders.R
\name{nan_handling_encoder_step}
\alias{nan_handling_encoder_step}
\title{Replace NaNs with 0 and append an is-nan indicator channel}
\usage{
nan_handling_encoder_step()
}
\description{
Replace NaNs with 0 and append an is-nan indicator channel
}
\keyword{internal}
