% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{predict_proba}
\alias{predict_proba}
\title{Predicted class probabilities}
\usage{
predict_proba(object, newdata, ...)
}
\arguments{
\item{object}{A fitted \code{tabfound_classifier}.}

\item{newdata}{Test features.}

\item{...}{Passed on.}
}
\description{
Shorthand for \code{predict(object, newdata, type = "prob")}.
}
