% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{predict_quantiles}
\alias{predict_quantiles}
\title{Predicted quantiles}
\usage{
predict_quantiles(object, newdata, quantiles = c(0.1, 0.5, 0.9), ...)
}
\arguments{
\item{object}{A fitted \code{tabfound_regressor}.}

\item{newdata}{Test features.}

\item{quantiles}{Numeric vector in \verb{(0, 1)}.}

\item{...}{Passed on.}
}
\description{
Shorthand for \code{predict(object, newdata, type = "quantiles")}.
}
