% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-quantile.R
\name{quantile_dist_icdf}
\alias{quantile_dist_icdf}
\title{Evaluate the quantile function of a fitted quantile distribution}
\usage{
quantile_dist_icdf(dist, probs)
}
\arguments{
\item{dist}{Result of \code{\link[=quantile_dist]{quantile_dist()}}.}

\item{probs}{Numeric vector of probabilities.}
}
\value{
A matrix \verb{(n_rows, length(probs))}.
}
\description{
Evaluate the quantile function of a fitted quantile distribution
}
\keyword{internal}
