% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep-transforms.R
\name{remove_constant_features_fit}
\alias{remove_constant_features_fit}
\title{Drop columns that are constant across all training rows.}
\usage{
remove_constant_features_fit(X_train)
}
\description{
Drop columns that are constant across all training rows.
}
\keyword{internal}
