% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-utils.R
\name{require_suggested}
\alias{require_suggested}
\title{Check that a suggested package is installed, abort with an install hint if not}
\usage{
require_suggested(pkg)
}
\description{
Check that a suggested package is installed, abort with an install hint if not
}
\keyword{internal}
