% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-utils.R
\name{set_names}
\alias{set_names}
\title{Construct a named character vector allowing duplicate names}
\usage{
set_names(x, nms)
}
\description{
\code{stats::setNames()} is fine for this, but cli's bullet lists need
repeated \code{"*"} names, which reads more clearly with a helper.
}
\keyword{internal}
