% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn-blocks.R
\name{set_transformer}
\alias{set_transformer}
\title{Stack of induced self-attention blocks}
\usage{
set_transformer(num_blocks, embedding_dim, n_heads, dim_ff, num_inds)
}
\description{
Stack of induced self-attention blocks
}
\keyword{internal}
