% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-tabicl-layers.R
\name{tabicl_set_transformer}
\alias{tabicl_set_transformer}
\title{Stack of induced self-attention blocks}
\usage{
tabicl_set_transformer(
  num_blocks,
  embedding_dim,
  n_heads,
  dim_ff,
  num_inds,
  ssmax = FALSE,
  bias_free_ln = FALSE,
  activation = "gelu"
)
}
\description{
Stack of induced self-attention blocks
}
\keyword{internal}
