% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-tabpfn-layers.R
\name{tabpfn_layer_stack}
\alias{tabpfn_layer_stack}
\title{Layer stack}
\usage{
tabpfn_layer_stack(
  n_layers,
  embedding_dim,
  n_heads,
  mlp_hidden_dim,
  layer_norm_eps = 1e-05,
  activation = "gelu"
)
}
\description{
Layer stack
}
\keyword{internal}
