% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-tabpfn-inputprep.R
\name{torch_nanmean_sb}
\alias{torch_nanmean_sb}
\title{Column-wise mean ignoring NaN (counts +Inf/-Inf as NaN too)}
\usage{
torch_nanmean_sb(x, include_inf = TRUE)
}
\description{
Column-wise mean ignoring NaN (counts +Inf/-Inf as NaN too)
}
\keyword{internal}
