% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-tabpfn-inputprep.R
\name{torch_nanstd_sb}
\alias{torch_nanstd_sb}
\title{Column-wise std ignoring NaN/Inf (unbiased, like torch.std default)}
\usage{
torch_nanstd_sb(x)
}
\description{
Column-wise std ignoring NaN/Inf (unbiased, like torch.std default)
}
\keyword{internal}
