% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep-transforms.R
\name{transform_squashing_scaler}
\alias{transform_squashing_scaler}
\title{Apply a fitted SquashingScaler to data.}
\usage{
transform_squashing_scaler(X, fit)
}
\description{
Apply a fitted SquashingScaler to data.
}
\keyword{internal}
