% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep-transforms.R
\name{transform_svd_features}
\alias{transform_svd_features}
\title{Apply a fitted SVD transformer.}
\usage{
transform_svd_features(X, fit)
}
\description{
Apply a fitted SVD transformer.
}
\keyword{internal}
