% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-tabpfn-encoders.R
\name{y_linear_step}
\alias{y_linear_step}
\title{y encoder's final linear step: \verb{y_encoder.*.layer.*}}
\usage{
y_linear_step(in_channels, embedding_dim, bias = TRUE)
}
\description{
y encoder's final linear step: \verb{y_encoder.*.layer.*}
}
\keyword{internal}
