% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep-transforms.R
\name{yeojohnson_forward}
\alias{yeojohnson_forward}
\title{Forward Yeo-Johnson power transform, given \code{lambda}.}
\usage{
yeojohnson_forward(x, lambda)
}
\description{
Forward Yeo-Johnson power transform, given \code{lambda}.
}
\keyword{internal}
